+++ src/CommandLineUtilities/ProgramStatus.cxx
@@ -176,9 +176,9 @@ class ProgramStatus : public Program
         }
       }
     } else if (cardType == CardType::type::Cru) {
-      formatHeader = "  %-9s %-16s %-10s %-14s %-15s %-10s %-14s %-14s %-8s %-19s %-11s %-7s\n";
-      formatRow = "  %-9s %-16s %-10s %-14s %-15s %-10s %-14.2f %-14.2f %-8s %-19.1f %-11s %-7s\n";
-      header = (boost::format(formatHeader) % "Link ID" % "GBT Mode Tx/Rx" % "Loopback" % "GBT MUX" % "Datapath Mode" % "Datapath" % "RX freq(MHz)" % "TX freq(MHz)" % "Status" % "Optical power(uW)" % "System ID" % "FEE ID").str();
+      formatHeader = "  %-9s %-16s %-10s %-14s %-15s %-10s %-14s %-14s %-8s %-19s %-11s %-7s %-10s\n";
+      formatRow = "  %-9s %-16s %-10s %-14s %-15s %-10s %-14.2f %-14.2f %-8s %-19.1f %-11s %-7s %-10s\n";
+      header = (boost::format(formatHeader) % "Link ID" % "GBT Mode Tx/Rx" % "Loopback" % "GBT MUX" % "Datapath Mode" % "Datapath" % "RX freq(MHz)" % "TX freq(MHz)" % "Status" % "Optical power(uW)" % "System ID" % "FEE ID" % "Glitches").str();
       lineFat = std::string(header.length(), '=') + '\n';
       lineThin = std::string(header.length(), '-') + '\n';
 
@@ -379,6 +379,7 @@ class ProgramStatus : public Program
 
         float rxFreq = link.rxFreq;
         float txFreq = link.txFreq;
+        int glitchCounter = link.glitchCounter;
 
         std::string linkStatus;
         if (link.stickyBit == Cru::LinkStatus::Up) {
@@ -407,6 +408,7 @@ class ProgramStatus : public Program
                              .addValue(opticalPower, "opticalPower")
                              .addValue(systemId, "systemId")
                              .addValue(feeId, "feeId")
+                             .addValue(glitchCounter, "glitchCounter")
                              .addTag(tags::Key::SerialId, card.serialId.getSerial())
                              .addTag(tags::Key::Endpoint, card.serialId.getEndpoint())
                              .addTag(tags::Key::CRU, card.sequenceId)
@@ -427,11 +429,12 @@ class ProgramStatus : public Program
           linkNode.put("opticalPower", Utilities::toPreciseString(opticalPower));
           linkNode.put("systemId", systemId);
           linkNode.put("feeId", feeId);
+          linkNode.put("glitchCounter", glitchCounter);
 
           // add the link node to the tree
           root.add_child(std::to_string(globalId), linkNode);
         } else {
-          auto format = boost::format(formatRow) % globalId % gbtTxRxMode % loopback % gbtMux % datapathMode % enabled % rxFreq % txFreq % linkStatus % opticalPower % systemId % feeId;
+          auto format = boost::format(formatRow) % globalId % gbtTxRxMode % loopback % gbtMux % datapathMode % enabled % rxFreq % txFreq % linkStatus % opticalPower % systemId % feeId % glitchCounter;
           table << format;
         }
       }