+++ src/Cru/CruBar.cxx
@@ -503,6 +503,7 @@ Cru::ReportInfo CruBar::report(bool forConfig)
     link.stickyBit = gbt.getStickyBit(link);
     link.rxFreq = gbt.getRxClockFrequency(link) / 1e6; // Hz -> Mhz
     link.txFreq = gbt.getTxClockFrequency(link) / 1e6; // Hz -> Mhz
+    link.glitchCounter = gbt.getGlitchCounter(link);
     link.systemId = datapathWrapper.getSystemId(link);
     link.feeId = datapathWrapper.getFeeId(link);