+++ src/Cru/Gbt.h
@@ -52,6 +52,7 @@ class Gbt
   uint32_t getRxClockFrequency(Link link);
   uint32_t getTxClockFrequency(Link link);
   uint32_t getGlitchCounter(Link link);
+  uint32_t getFecCounter(Link link);
   void resetFifo();
   std::map<int, LoopbackStats> getLoopbackStats(bool reset, GbtPatternMode::type patternMode = GbtPatternMode::type::Counter,
                                                 GbtCounterType::type counterType = GbtCounterType::type::ThirtyBit,