+++ doc/releaseNotes.md
@@ -24,3 +24,6 @@ This file describes the main feature changes for released versions of ReadoutCar
 
 ## v0.40.2 - 16/02/2023
 - Fixed bug with o2-roc-ctp-emulator failing to parse hexadecimal values of the init-orbit option.
+
+## next version
+- Added glitchCounter to roc-status (all output styles, including monitoring).