+++ doc/releaseNotes.md
@@ -28,3 +28,4 @@ This file describes the main feature changes for released versions of ReadoutCar
 ## next version
 - Added glitchCounter to roc-status (monitoring + JSON output).
 - Added DMA status (enabled / disabled) to roc-status (all output styles, including monitoring).
+- Added FEC counter per link to roc-status (stdout + JSON output).